requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
Pillow==9.3.0
pony==0.7.16
psycopg2==2.9.5
pyTelegramBotAPI==4.7.1
requests==2.28.1
telebot==0.0.4
urllib3==1.26.12
